requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

transformers==4.44.2
torch==2.3.1
pandas==2.1.2
scikit-learn==1.2.2
azure-storage-blob==12.23.1
azure-identity==1.18.0
openai==1.45.0
aio-pika==9.4.3
SQLAlchemy==2.0.35
pyodbc==5.1.0